feat: support logo colour inversion in dark mode
